Rework the file writing logic
This will hopefully avoid running out of memory when writing a large image. It takes almost the double of the time to save an image, as the script does now save a blank image file and then gradually overwrite the pixel data of the bitmap. I also added the draw_dot function.