Fix a bug that caused script crash
It happened when there were more than one alt action, of which the first one had no states, but the other(s) did. The script so happened to try to read from an uninitialized matrix value, which causes Lua scripts to crash, since the core doesn't know how to deal with a script that tries to index an invalid variable.